webapp/pyproject.toml
[project]
name = "app-precos-web"
version = "0.1.0"
description = "Django PWA that mirrors the AppPrecos Android client with Supabase backend."
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "httpx>=0.28.1",
    "pillow>=12.0.0",
    "playwright>=1.56.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.24.0",
    "whitenoise>=6.8.2",
]
